Test.Tasty.QuickCheck

Description

This module allows to use QuickCheck properties in tasty.

Documentation

testProperty :: Testable a => TestName -> a -> TestTree

Create a Test for a QuickCheck Testable property

newtype QuickCheckTests

Number of test cases for QuickCheck to generate

Constructors

QuickCheckTests Int

newtype QuickCheckReplay

Replay a previous test using a replay token

Constructors

QuickCheckReplay (Maybe (QCGen, Int))

newtype QuickCheckShowReplay

If a test case fails unexpectedly, show the replay token

Constructors

QuickCheckShowReplay Bool

newtype QuickCheckMaxSize

Size of the biggest test cases

Constructors

QuickCheckMaxSize Int

newtype QuickCheckMaxRatio

Maximum number of of discarded tests per successful test before giving up.

Constructors

QuickCheckMaxRatio Int

newtype QuickCheckVerbose

Show the test cases that QuickCheck generates

Constructors

QuickCheckVerbose Bool

module Test.QuickCheck

Internal

This is exposed for testing purposes and not considered as a part of the public API. You probably shouldn't need it.

newtype QC

Constructors

QC Property